🐛 fix: sanitize DeepSeek surrogate payloads
